cleanup the makePot main file from utility functions (moving everithing to fs/fs)
